Moved card.tsx
