+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/card/index.tsx
